Fix dtype bug in test

+++ tests/test_tools.py
@@ -107,7 +107,7 @@ class TestDownsampleUpsample(object):
         filt = np.zeros([size, size + 1])
         filt[5, 5] = 1
         filt = scipy.ndimage.gaussian_filter(filt, sigma=1)
-        filt = torch.tensor(filt, dtype=torch.float32, device=DEVICE)
+        filt = torch.tensor(filt, dtype=img.dtype, device=DEVICE)
         img_down = po.tools.correlate_downsample(img, filt=filt)
         img_up = po.tools.upsample_convolve(img_down, odd=(odd, 1), filt=filt)
         assert np.unravel_index(img_up.cpu().numpy().argmax(), img_up.shape) == (0, 0, 12, 12)
